requirements.txt
# Core Algorand SDK for Python
py-algorand-sdk>=2.0.0

# PyTeal for writing smart contracts in Python
pyteal>=0.24.0

# Environment variable management
python-dotenv>=1.0.0

# Optional: for better CLI output
colorama>=0.4.6
